Implement Color Coding
