+++ src/components/ListPersonComponent.jsx
@@ -3,6 +3,7 @@ import PersonService from '../services/PersonService';
 import '../App.css';
 
 import { Table, Button, ButtonGroup, ButtonToolbar, Modal, Form, InputGroup, Alert } from "react-bootstrap";
+import RowComponent from './RowComponent';
 
 const ListPersonComponent = () => {  
     //Database structure
@@ -18,29 +19,9 @@ const ListPersonComponent = () => {
     //import
     const [file, setFile] = useState(null);
 
-    // more info for students
-    const [modalData, setModalData] = useState({});
-    const [counterpartData, setCounterpartData] = useState({});
-    const [isGuest, setIsGuest] = useState(false);
-    const [show, setShow] = useState(false);
+    //color code table
     
     const closeDelete = () => setClearDatabase(false);
-
-    const handleClose = () => setShow(false);
-    
-    const handleShow = (person) => {
-        setModalData(person);
-        
-        setIsGuest(people.find(o => o.guestTicket === person.ticket) != null ? true : false);
-        
-        if(person.guest === 'Y'){
-            setCounterpartData(people.find(o => o.ticket === person.guestTicket));
-        } else if (isGuest) {
-            setCounterpartData(people.find(o => o.guestTicket === person.ticket));
-        }
-
-        setShow(true);        
-    };
     
     //redirect input to search bar
     function inputFocus(){
@@ -56,7 +37,7 @@ const ListPersonComponent = () => {
             }
         )
     }, [people]);
-    
+
     return (
         <div>
             <h2 className = "text-center" style={{ color: "white" }}> Student List </h2>
@@ -187,116 +168,13 @@ const ListPersonComponent = () => {
                                 }
                             }).map(
                                 person =>
-                                <tr key = {person.id} style = {{backgroundColor: person.checkedIn == null ? null : "red"}}>
-                                    <td> {person.ticket} </td>
-                                    <td> {person.countyId} </td>
-                                    <td> {person.lastName} </td>
-                                    <td> {person.middleInitial} </td>
-                                    <td> {person.firstName} </td>
-                                    <td> {person.grade} </td>
-                                    <td> {person.paymentMethod} </td>
-                                    <td> {person.guest} </td>
-                                    <td> {person.guestTicket} </td>
-                                    <td> 
-                                        <ButtonGroup role = "group" size = 'sm' >
-                                            <Button variant = "dark" onClick = {() => handleShow(person) }>More Info</Button>
-                                            <Button variant = "dark" onClick = {() => {
-                                                console.log(person.checkedIn);
-                                                if(person.checkedIn === null) {
-                                                    person.checkedIn = true;
-                                                } else{
-                                                    person.checkedIn ? person.checkedIn = false : person.checkedIn = true;
-                                                    console.log(person.checkedIn);
-                                                }
-                                                }}
-                                            >Check In</Button>
-                                        </ButtonGroup>
-                                    </td>
-                                </tr>
+                                <RowComponent person = {person} people = {people} />
                             )
                         }
                     </tbody>
                 </Table>
 
-            <Modal
-                size="lg"
-                show={show} 
-                onHide={handleClose}
-            >
-            <Modal.Header>
-                <Modal.Title>More Info on {modalData.firstName + " " + modalData.lastName}</Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    <h3>{ !isGuest ? "Student: " : "Guest: " }</h3>
-                    <table className="table">
-                        <thead>
-                            <tr>
-                                <th>Ticket</th>
-                                <th>ID</th>
-                                <th>LAST</th>
-                                <th>MI</th>
-                                <th>FIRST</th>
-                                <th>GR</th>
-                                <th>Payment Method</th>
-                                <th>Guest YN</th>
-                                <th>Guest Ticket Number</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td> {modalData.ticket} </td>
-                                <td> {modalData.countyId} </td>
-                                <td> {modalData.lastName} </td>
-                                <td> {modalData.middleInitial} </td>
-                                <td> {modalData.firstName} </td>
-                                <td> {modalData.grade} </td>
-                                <td> {modalData.paymentMethod} </td>
-                                <td> {modalData.guest} </td>
-                                <td> {modalData.guestTicket} </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                   { (modalData.guest === 'Y' || isGuest) &&
-                    <>
-                        <h3>{ isGuest ? "Student: " : "Guest: " } </h3>
-                        <table className="table">
-                            <thead>
-                                <tr>
-                                    <th>Ticket</th>
-                                    <th>ID</th>
-                                    <th>LAST</th>
-                                    <th>MI</th>
-                                    <th>FIRST</th>
-                                    <th>GR</th>
-                                    <th>Payment Method</th>
-                                    <th>Guest YN</th>
-                                    <th>Guest Ticket Number</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <tr>
-                                <td> {counterpartData.ticket} </td>
-                                <td> {counterpartData.countyId} </td>
-                                <td> {counterpartData.lastName} </td>
-                                <td> {counterpartData.middleInitial} </td>
-                                <td> {counterpartData.firstName} </td>
-                                <td> {counterpartData.grade} </td>
-                                <td> {counterpartData.paymentMethod} </td>
-                                <td> {counterpartData.guest} </td>
-                                <td> {counterpartData.guestTicket} </td>
-                                </tr>
-                            </tbody>
-                        </table>
-                    </>
-                   }
-                    <h3><a href=''>Logs</a></h3>
-                </Modal.Body>
-                <Modal.Footer>
-                    <Button variant="secondary" onClick={handleClose}>
-                        Close
-                    </Button>
-                </Modal.Footer>
-            </Modal>
+            
         </div>
     );
 };
